pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "videothinkbench"
version = "0.1.0"
description = "VideoThinkBench unified data generation and evaluation toolkit"
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.24.0",
    "opencv-python>=4.8.0",
    "Pillow>=10.0.0",
    "requests>=2.31.0",
    "tqdm>=4.66.0",
]

[project.scripts]
videothinkbench = "cli:main"

[tool.setuptools]
include-package-data = true
py-modules = ["cli"]

[tool.setuptools.packages.find]
include = [
    "data*",
    "evaluation*",
    "training*",
    "core*",
]
namespaces = false

[tool.setuptools.package-data]
data.visioncentric.visual_puzzles = ["fonts/*", "example_data/*/*.json", "example_data/*/*.png"]
evaluation.textcentric = ["config/*.yaml"]
